requirements.txt
Django~=4.2.7
django-debug-toolbar~=4.2.0
django-bootstrap-v5~=1.0.11
django-bootstrap-icons~=0.8.6
django-jquery~=3.1.0
django-bootstrap-datepicker-plus~=5.0.4
django-crispy-forms~=2.0
crispy-bootstrap5~=2023.10
djangorestframework~=3.14.0
django-cors-headers~=4.3.1